Move user event params to new class. Rename enum paramName to text to have common name.
