+++ RingPublishingTracking/src/androidTest/java/com/ringpublishing/tracking/internal/decorator/UserSsoDataDecoratorTest.kt
@@ -29,7 +29,7 @@ class UserSsoDataDecoratorTest
 		val event = Event()
 		userSsoDataDecorator.decorate(event)
 
-		val result = event.parameters[EventParam.USER_SSO_DATA.paramName] as String?
+		val result = event.parameters[EventParam.USER_SSO_DATA.text] as String?
 
 		val decodedResult = String(Base64.decode(result, Base64.DEFAULT))
