+++ RingPublishingTracking/src/main/java/com/ringpublishing/tracking/internal/factory/UserEventParam.kt
@@ -1,0 +1,13 @@
+/*
+ *  Created by Grzegorz Małopolski on 10/11/21, 4:27 PM
+ * Copyright © 2021 Ringier Axel Springer Tech. All rights reserved.
+ *
+ */
+
+package com.ringpublishing.tracking.internal.factory
+
+internal enum class UserEventParam(val text: String)
+{
+	SELECTED_ELEMENT_NAME("VE"),
+	TARGET_URL("VU")
+}
